+++ src/MagicOnion.Server.JsonTranscoding.Swagger/MagicOnionJsonTranscodingErrorResponse.cs
@@ -1,0 +1,3 @@
+namespace MagicOnion.Server.JsonTranscoding.Swagger;
+
+internal record MagicOnionJsonTranscodingErrorResponse(int Code, string Detail);
